requirements/common.txt
Babel==2.4.0
Flask==0.12
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-Caching==1.2.0
Flask-Jsonpify==1.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Restless==0.17.0
flask-rq2==17.0
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
-e git://github.com/spacedogXYZ/Flask-Store.git#egg=flask-store
Flask-WTF==0.14.2
Jinja2==2.9.6
Mako==1.0.1
MarkupSafe==1.0
PyYAML==3.12
SQLAlchemy==1.1.11
SQLAlchemy-Utils==0.32.14
WTForms==2.1
WTForms-Components==0.9.7
Werkzeug==0.10.4
alembic==0.8.8
blinker==1.3
decorator==3.4.2
geopy==1.11.0
httplib2==0.8
infinity==1.3
intervals==0.3.1
itsdangerous==0.24
mimerender==0.5.5
phonenumbers==7.0.4
-e git://github.com/ncadou/pyrepresent.git#egg=pyrepresent
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2017.2
requests==2.2.1
speaklater==1.3
pyopenstates==1.1.0
twilio==6.4.2
unittest2==0.5.1
validators==0.7
webassets==0.12.1
wsgiref==0.1.2
